refactor(diffusion): consolidate test utils, remove helpers.js
Move detectPlatform, setupJsLogger, isPng into utils.js and update
generate-image.test.js to import from utils.js only. Add platform
detection for device selection in model-loading.test.js.
