+++ packages/lib-infer-diffusion/test/integration/model-loading.test.js
@@ -2,9 +2,16 @@
 
 const test = require('brittle')
 const FilesystemDL = require('@qvac/dl-filesystem')
+const os = require('bare-os')
 
 const ImgStableDiffusion = require('../../index.js')
 const { ensureModel } = require('./utils')
+
+const platform = os.platform()
+const arch = os.arch()
+const isDarwinX64 = platform === 'darwin' && arch === 'x64'
+const isLinuxArm64 = platform === 'linux' && arch === 'arm64'
+const useCpu = isDarwinX64 || isLinuxArm64
 
 const DEFAULT_MODEL = {
   name: 'stable-diffusion-v2-1-Q8_0.gguf',
@@ -20,7 +27,7 @@ test('model loading - load and unload', { timeout: 600_000 }, async t => {
   const loader = new FilesystemDL({ dirPath: modelDir })
   const config = {
     threads: '4',
-    device: 'gpu',
+    device: useCpu ? 'cpu' : 'gpu',
     prediction: 'v'
   }
 
@@ -46,6 +53,8 @@ test('model loading - load and unload', { timeout: 600_000 }, async t => {
 })
 
 // Keep event loop alive briefly to let pending async operations complete
+// This prevents C++ destructors from running while async cleanup is still happening
+// which can cause segfaults (exit code 139)
 setImmediate(() => {
   setTimeout(() => {}, 500)
 })
